--- src/app/api/users/route.ts
@@ -1,1 +1,0 @@
-
