+++ src/app/layout.tsx
@@ -18,7 +18,7 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
-      <body className={inter.className}>
+      <body className={inter.className} suppressHydrationWarning={true}>
         <AnimatedBackground />
         <div className="content-wrapper">
           <Providers>
