+++ src/app/payment/momo-return/page.tsx
@@ -1,12 +1,12 @@
 'use client';
 
-import { useEffect, useState } from 'react';
+import { useEffect, useState, Suspense } from 'react';
 import { useRouter, useSearchParams } from 'next/navigation';
 import { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from '@/components/ui/card';
 import { Button } from '@/components/ui/button';
 import { Loader2 } from 'lucide-react';
 
-export default function MomoReturn() {
+function MomoReturnContent() {
   const [status, setStatus] = useState<'loading' | 'success' | 'error'>('loading');
   const [message, setMessage] = useState('Đang xử lý kết quả thanh toán...');
   const router = useRouter();
@@ -76,5 +76,22 @@ export default function MomoReturn() {
         </CardFooter>
       </Card>
     </div>
+  );
+}
+
+export default function MomoReturn() {
+  return (
+    <Suspense fallback={
+      <div className="container mx-auto py-10 flex justify-center items-center min-h-[80vh]">
+        <Card className="w-full max-w-md bg-white/95 backdrop-blur-xl rounded-3xl shadow-xl border-0">
+          <CardContent className="flex flex-col items-center space-y-4 py-8">
+            <Loader2 className="h-16 w-16 text-[#7F886A] animate-spin" />
+            <p className="text-center text-gray-700">Đang tải...</p>
+          </CardContent>
+        </Card>
+      </div>
+    }>
+      <MomoReturnContent />
+    </Suspense>
   );
 }
